Add files via upload
We wrote the besselCompute function to handle the particle's interactions inside and outside the light cone; had to make B an array because of the way the integral function works. Goal now is to fiddle with timespan, particle path, etc. (especially try giving the particle very rough oscillation - close to speed of light)